MB-66210: Point and MultiPoint changes
 - Switched out all cell operations to point operations
 - All point comparisions will use approx equal which has an epsilon of 1e-15
 - LineStrings and MultiLineStrings will project point onto themselves
and compare the distance between points for intersection
 - Polygons and MultiPolygons use shape indexes with a Closed Vertex Model
which means the boundaries and the vertices are considered part of the shape
 - Circle will use distance from center to identify intersecting points
 - Envelopes will use bound checks to identify intersects
 - New Circles and Envelopes will now init themselves before returning
 - Added appropriate test cases

Note:
    Projection onto a segment calculation might not be completely accurate.
Projection of point (1,1) onto the line (-2,1) amd (2,1) returns a point that
has a distance of 50m from (1,1). Unable to verify if this distance is actually
accurate or not.