MB-66210: LineString and MultiLineString changes (#19)
- Changed polylines intersects polygons function to first make a shape
index of the polygons and then perform a series of contains point
queries to identify if any of the vertices fall inside the polygon. If
this doesn't yield a result, it means that there is an edge crossing
with no vertex overlap, so we perform all edge intersect queries
 - Removed unnecessary loop within edge crossing logic
- Changed polyline intersects envelope by checking each vertex's bounds
and then doing an edge crossing check
 - Removed unecessary loop within this edge crossing logic as well
- Fixed polyline intersects circle logic. Previously would only check a
single edge of the polyline before exiting the function
- Added contains logic for points and multipoints as they are very
similar to intersects.
 - Added appropriate test cases covering edge cases